Move the docker_image library to a custom resource
